Add files via upload
The intent of this program segment is two-fold:
     1 - create a new file from the original .csv (so I can still use all the data)   
     2 - strip the unneeded columns off the data (I only need closing numbers)

I included a few .csv files for eval.  These data files are now downloaded via an API  program and stored in the local folder with this "column removal tool"

The next step is  to combine ~2,500 of  these files into a full matrix, to be used for data analysis.